Restructured: hooks folder with individual file
