+++ src/hooks/useOutsideClick.ts
@@ -1,0 +1,33 @@
+import * as React from "react";
+
+export const useOutsideClick = (
+  elementRef: React.RefObject<HTMLElement>,
+  callback: (event: MouseEvent) => void
+) => {
+  const callbackRef = React.useRef<(e: MouseEvent) => void>();
+
+  if (!callbackRef.current) {
+    callbackRef.current = callback;
+  }
+
+  React.useEffect(() => {
+    const handleOutsideClick = (e: MouseEvent) => {
+      if (
+        !elementRef?.current?.contains(e.target as Element) &&
+        callbackRef.current
+      ) {
+        callbackRef.current(e);
+      }
+    };
+
+    if (callbackRef.current) {
+      document.addEventListener("click", handleOutsideClick, true);
+    }
+
+    return () => {
+      if (callbackRef.current) {
+        document.addEventListener("click", handleOutsideClick, true);
+      }
+    };
+  }, [callbackRef.current, elementRef]);
+};
